Completes Step 4 (API) and Prepares for Step 5
Step 1: GitHub Action

Corrected the action to search for test_ml.py
in the tests/ directory that I added earlier
in the project.

Step 4: Pytest

Added two new Pytests and two new Pytest fixtures
to perform testing of the GET and POST endpoints
during the Pytest run.

Step 4: API

Implemented the GET endpoint at the root of the
application. Returns a single string.

Implemented the POST endpoint at /data/ to take
in one set of features and return an inference
using the trained model, OneHotEncoder, and
StandardScaler from disk.